pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "stockbot"
version = "0.1.0"
description = "AI-powered autonomous stock trading bot with LangGraph agents"
requires-python = ">=3.11"
dependencies = [
    # Broker
    "alpaca-py>=0.30,<1.0",

    # AI / Agents
    "langgraph>=0.3,<1.0",
    "langchain>=0.3,<1.0",
    "langchain-anthropic>=0.3,<1.0",
    "langchain-openai>=0.3,<1.0",

    # Data & Analysis
    "pandas>=2.2,<3.0",
    "numpy>=1.26,<3.0",
    "pandas-ta>=0.3,<1.0",

    # Database
    "duckdb>=1.1,<2.0",
    "sqlmodel>=0.0.22,<1.0",

    # Dashboard
    "streamlit>=1.40,<2.0",
    "plotly>=5.24,<6.0",

    # Scheduling & Market Calendar
    "apscheduler>=3.10,<4.0",
    "exchange-calendars>=4.5,<5.0",

    # Configuration
    "pydantic-settings>=2.6,<3.0",

    # Logging
    "structlog>=24.4,<26.0",

    # HTTP & Async
    "httpx>=0.28,<1.0",

    # Report generation
    "jinja2>=3.1,<4.0",

    # CLI
    "typer>=0.14,<1.0",
    "rich>=13.9,<15.0",

    # Utilities
    "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "pytest-mock>=3.14,<4.0",
    "ruff>=0.8,<1.0",
]

[tool.setuptools.packages.find]
include = ["stockbot*", "config*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
